% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_countplot.R
\name{aa_countplot}
\alias{aa_countplot}
\title{Count and plot amino aicd occurence}
\usage{
aa_countplot(..., plot = FALSE)
}
\arguments{
\item{...}{Protein sequence as a character string.}

\item{plot}{A \code{logical} argument to decide whether to plot or not the resuls.}
}
\value{
dataframe with 3 columns: name of a sequence, its amino acid residues and their count.
a plot if \code{plot} argument is set to \code{TRUE} .
}
\description{
Count the occurence of amino aicd residues in a protein sequence and plot their relative percentages.
}
\examples{
H3_seq <-  "ARTKQTARKSTGGKAPRKQLATKAARKSAPATGGVKKPHRYRPGTVALREIRRYQKSTELLIRKLPFQRLVREIAQDFKTDLRFQSSAVMALQEACEAYLVGLFEDTNLCAIHAKRVTIMPKDIQLARRIRGERA"
H4_seq <-  "SGRGKGGKGLGKGGAKRHRKVLRDNIQGITKPAIRRLARRGGVKRISGLIYEETRGVLKVFLENVIRDAVTYTEHAKRKTVTAMDVVYALKRQGRTLYGFGG"
aa_countplot(H3_seq)
aa_countplot(H3_seq, H4_seq, plot = TRUE)
}
