% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf_searchDiagIon.R
\name{diagIon_find}
\alias{diagIon_find}
\title{Find MS/MS spectra that contain a diagnostic ion.}
\usage{
diagIon_find(mgf_file, diag_ion, tol = 0.01)
}
\arguments{
\item{mgf_file}{mgf file to search.}

\item{diag_ion}{The mz of the diagnsotic ion}

\item{tol}{A mass tolerance to respect during the search}
}
\value{
A \code{tibble} with 6 columns.
}
\description{
Find MS/MS spectra that contain a diagnostic ion.
}
