% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_extractMetaData.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract raw file and search file names}
\usage{
extract_metadata(filename, save_file = FALSE)
}
\arguments{
\item{filename}{Proteomics experiment excel output.}
}
\value{
\code{.csv} file with 3 columns: \code{file}, \code{channel}, and \code{dat}.
}
\description{
Extract .raw file and .dat file names from Proline output excel file (\code{"Search settings and infos"} sheet).
}
