% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histonePTM-package.R
\docType{package}
\name{histonePTM-package}
\alias{histonePTM}
\alias{histonePTM-package}
\title{histonePTM: A toolset to analyze histone PTMs proteomics data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Handy functions to organize, analyse, and visualize histone PTMs proteomics data and "what you have".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HijaziHassan/histonePTM}
  \item Report bugs at \url{https://github.com/HijaziHassan/histonePTM/issues}
}

}
\author{
\strong{Maintainer}: Hassan Hijazi \email{hejaziihassan@gmail.com} (\href{https://orcid.org/0000-0001-6408-1768}{ORCID})

}
\keyword{internal}
