% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-histptm_lookup.R
\docType{data}
\name{histptm_lookup}
\alias{histptm_lookup}
\title{Histone PTM lookup vector}
\format{
a named character vector.
}
\source{
in-house user defined names
}
\usage{
histptm_lookup
}
\description{
a named vector containing the PTM annotations and their corresponding shorthand names
}
\keyword{datasets}
