% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-histptm_mass.R
\docType{data}
\name{histptm_mass}
\alias{histptm_mass}
\title{Histone PTM delta masses}
\format{
a named character vector.
}
\source{
Unimod
}
\usage{
histptm_mass
}
\description{
a named vector containing the PTM annotations and their corresponding mass shifts
}
\keyword{datasets}
