% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-histptm_unimod.R
\docType{data}
\name{histptm_unimod}
\alias{histptm_unimod}
\title{Histone PTM Unimod accession number}
\format{
a named character vector.
}
\source{
Unimod
}
\usage{
histptm_unimod
}
\description{
a named vector containing the PTM annotations and their corresponding unimod acession number.
}
\keyword{datasets}
