% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litReview.R
\name{litReview}
\alias{litReview}
\title{A quick literature review on NCBI}
\usage{
litReview(
  start,
  end,
  term,
  db = "pubmed",
  save_plot = FALSE,
  save_file = FALSE
)
}
\arguments{
\item{start}{An \code{integer} specifying the year from which to start the search.}

\item{end}{An \code{integer} specifying the year at which to stop the search.}

\item{term}{any search term including boolean operators (i.e. AND, NOT,OR, ...) and tags ([au] for author,
[lau] for last author, review [pt] for publication type, [ti] for title etc...Check \href{https://pubmed.ncbi.nlm.nih.gov/help/#using-search-field-tags}{here}
for an exhaustive list.}

\item{db}{database to searhc in. defaults is "pubmed". check \code{entrez_dbs()} for all options.}

\item{save_plot}{\code{bool}. If `TRUE` a line plot of count vs year colored by search term will be exported as `png`.}

\item{save_file}{\code{bool}. If `TRUE` results will be saved as `.csv` files.}
}
\value{
A list. A dataframe with  year, id, and article title. A vector summarizing the article counts per year and a a line plot of it.
}
\description{
A wrapper function around \code{rentrez} package functions to extract article titles using year and search term.
Check \code{rentrez package} \href{https://cran.r-project.org/web/packages/rentrez/vignettes/rentrez_tutorial.html}{documentation}.
}
\examples{
\dontrun{
litReview(start = 2019, end = 2024, term = "Huntington's disease
AND histone post-translational modifications")
}

}
