% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litRicher.R
\name{litRicher}
\alias{litRicher}
\title{Quotes from literature about histones}
\usage{
litRicher(quotes_num = 1)
}
\arguments{
\item{quotes_num}{Number of quotes to print out (\code{integer}. To read all quotes, use "All".)}
}
\value{
A character string containing a quote about histones and histone PTMs.
}
\description{
Print a quote(s) from cherry-picked from histone literature.
}
