% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf_extract_ms2scan.R
\name{mgf_extract_ms2scan}
\alias{mgf_extract_ms2scan}
\title{Extract mz & intensity of an MS2 scan from an mgf file}
\usage{
mgf_extract_ms2scan(mgf_file, scan)
}
\arguments{
\item{mgf_file}{.mgf file.}

\item{scan}{scan number containing ms2 of interest}
}
\value{
\code{dataframe} containing 4 columns: _mz_, _intensity_, _scan_, and _file_.
}
\description{
Extract mz & intensity of an MS2 scan from an mgf file
}
