% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf_extractScan.R
\name{mgf_extractScan}
\alias{mgf_extractScan}
\title{Extract user-defined Scan from MGF file.}
\usage{
mgf_extractScan(mgf_file, scan)
}
\arguments{
\item{mgf_file}{mgf file (path) name}

\item{scan}{\code{numeric}. scan number}
}
\value{
tibble with 4 colums: file name, scan number and their corresponfing mz and intensity values.
}
\description{
wrapper function around \code{mgf.read} from \code{spectrum_utils} library in Python.
}
