% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf_searchDiagIon.R
\name{mgf_searchDiagIon}
\alias{mgf_searchDiagIon}
\title{Find MS/MS spectra that contain a diagnostic ion.}
\usage{
mgf_searchDiagIon(mgf_file, diag_ion, tol = 0.002)
}
\arguments{
\item{mgf_file}{mgf file to search.}

\item{diag_ion}{The mz of the diagnsotic ion}

\item{tol}{A mass tolerance to respect during the search}
}
\value{
A \code{tibble} with 6 columns diagnostic ion and its intensity relative to the base peak.
}
\description{
Report user-defined diagnsotic ion(s) per MS/MS spectrum.
}
