% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_clearLabeling.R
\name{misc_clearLabeling}
\alias{misc_clearLabeling}
\title{Reverse PTM Labeling}
\usage{
misc_clearLabeling(
  ptm_string,
  rules = NULL,
  labeling = NULL,
  residue = c("keep", "remove")
)
}
\arguments{
\item{ptm_string}{PTM-containing string}

\item{rules}{A named vector containing \code{(labeled_ptm = unlabeled_ptm)}. Example: \code{c("pr" = "un")}.}

\item{labeling}{The labeling agent used. Use `PA` (propionic anhydride), `TMA` ( trimethylacetic anhydride), or `PIC_PA` (phenylisocynate + PA).
If none applies or you don't want to use the built-in rules, set it to 'none' and indicate your `rules`. Otherwise, it defaults to `NULL`, returning the input as is.
There are embedded rules for those reagents. If those don't match your case, define your own rules and `labeling`must not be `none`.}

\item{residue}{Specify wether to remove residue(capitlized letter followed by number). Either `keep` or `remove`.}
}
\value{
unlabeled string or the same string if labeling is `none`.
}
\description{
For final representation it is better to replace chemical labeling with biological state. For example, replace 'propionyl' with 'unmodified'
or 'butyryl' with 'me1' (since me1+pr = bu).
}
