% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_createFolder.R
\name{misc_createFolder}
\alias{misc_createFolder}
\title{Create Folder}
\usage{
misc_createFolder(foldername)
}
\arguments{
\item{foldername}{name of the folder with or without forward-slash (/). Only the name after the last slash will be used. Any extension will be removed.}
}
\value{
folder
}
\description{
Create a folder carrying the name of the analyzed file (.csv or .xlsx). If a path is provided, only what after the last slash is considered.
}
