% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_extractMetaData.R
\name{misc_extractMetaData}
\alias{misc_extractMetaData}
\title{Extract raw file and search file names}
\usage{
misc_extractMetaData(filename, save_file = "")
}
\arguments{
\item{filename}{Proline export excel filename ('\code{Search settings and infos}' sheet is hardcoded.)}

\item{save_file}{Provide a file name if you want to save the extracted metadata.}
}
\value{
\code{.csv} file with 3 columns: \code{file}, \code{channel}, and \code{dat}.
}
\description{
Extract .raw file and .dat file names from Proline output excel file (\code{"Search settings and infos"} sheet).
}
