% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_parseSpectrumtitle.R
\name{misc_parseSpectrumtitle}
\alias{misc_parseSpectrumtitle}
\title{Split spectrum_title column into scan and file columns}
\usage{
misc_parseSpectrumtitle(df, col, scan_prefix, file_prefix, sep = "", ...)
}
\arguments{
\item{df}{dataframe}

\item{col}{The column to be fragmented.}

\item{scan_prefix}{The pattern the preceeds the scan number (e.g. 'Index:' or "Scan:"). If there is a space before it make sure to include it.}

\item{file_prefix}{The pattern the preceeds the file name (e.g. 'raw_file:). If there is a space before it make sure to include it.}

\item{sep}{A separator as ";" or ":" (`optional`).}

\item{...}{to add more arguments that could be useful as this function is a wrapper around `separate_wider_regex` function.}
}
\value{
two columns: `id_scan` and `id_filename`
}
\description{
Split spectrum_title column into scan and file columns
}
