% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_renameCols.R
\name{misc_renameCols}
\alias{misc_renameCols}
\title{Rename columns in Proline Output}
\usage{
misc_renameCols(analysisfile, metafile)
}
\arguments{
\item{analysisfile}{Name of Proline output excel file}

\item{metafile}{Name of user-defined sample names (at least two columns: SampleName [user defined sample names] and file [raw files]).}
}
\value{
renamed dataframe
}
\description{
Rename columns "abundance_..." from Proline proteomics results output.
}
