% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_useCustomNames.R
\name{misc_useCustomNames}
\alias{misc_useCustomNames}
\title{Rename 'abundance_xxx' columns in Proline Export.}
\usage{
misc_useCustomNames(raw_df, user_df)
}
\arguments{
\item{raw_df}{Dataframe extracted using \code{misc_extractMetaData} function.}

\item{user_df}{Dataframe containing "SampleName" and file" columns (could also contain 'Condition', 'Bioreplicate' or 'TechReplicate' columns).}
}
\value{
named vector
}
\description{
Create custom names to replace dat file coded 'abundance' and 'psm count' headers in Proline export.
}
