% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mirrorSpectra.R
\name{plot_mirrorSpectra}
\alias{plot_mirrorSpectra}
\title{Mirror Plot}
\usage{
plot_mirrorSpectra(
  data,
  mz_col,
  intensity_col,
  grouping_col,
  top_spec,
  bottom_spec,
  title = "",
  norm_basepeak = TRUE,
  readfromfile = FALSE
)
}
\arguments{
\item{data}{A \code{dataframe} or a comma-separated (txt, csv) or excel (.xls, xlsx) file.}

\item{mz_col}{mz values column}

\item{intensity_col}{intensity values column}

\item{grouping_col}{Any column that serve as ID. Could be scan number, or file, ... Those will annotate the y-axis.}

\item{top_spec}{Value (\code{character}) from the \code{grouping col}. Ex: 'processed_spec".}

\item{bottom_spec}{Value (\code{character}) from the \code{grouping col}. Ex: 'raw_spec".}

\item{title}{Title of the plot.}

\item{norm_basepeak}{\code{logical}. If \code{TRUE} (default), intensities will be normalized to base peak intensity.}

\item{readfromfile}{\code{logical}. If \code{FALSE} (default), it means that the \code{data} argument is  a \code{dataframe} and
not a file name (\code{.csv}, \code{txt}, \code{xls} or \code{xlsx}.}
}
\value{
A mirror plot.
}
\description{
Mirror Plot
}
