% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scheduledPRM.R
\name{plot_scheduledPRM}
\alias{plot_scheduledPRM}
\title{plot_scheduledPRM}
\usage{
plot_scheduledPRM(
  df,
  rt_col,
  mz_col,
  rt_window,
  peak_width = 0.3,
  save_plot = FALSE
)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} having at least two columns: mz and rt.}

\item{rt_col}{retention time column (min)}

\item{mz_col}{mass-over-charge column}

\item{rt_window}{The retention time window in min (a 5-min window means +-2.5 min).
multiple values can be submitted as a vector (e.g \code{c(5, 7)}).}

\item{peak_width}{The width of the elution peak in min (Default is 0.3).}

\item{save_plot}{a \code{logical} value to either save \code{TRUE} or to only view the plot \code{FALSE}.
default value is \code{TRUE}.}
}
\value{
plot.
}
\description{
A function to plot scheduled PRM graph to check how many precursor ions are concurrently
monitored over the same time by varying the \code{rt_window} size.
}
\examples{
plot_scheduledPRM(mtcars, mpg, cyl, c(5, 10), save_plot = TRUE)

}
