% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_beautify.R
\name{ptm_beautify}
\alias{ptm_beautify}
\title{ptm_beautify}
\usage{
ptm_beautify(
  PTM,
  lookup,
  software = c("Proline", "Skyline"),
  residue = c("keep", "remove")
)
}
\arguments{
\item{PTM}{The PTM string to be shorthanded}

\item{lookup}{The shorthanded names to replace the long PTM names (`histptm_lookup` and `shorthistptm_mass` are named vector provided by default.
The user can provide any named vector adapted to the PTM being used.)}

\item{software}{The software used for proteomic analysis (ex. Proline (Default) or Skyline)}

\item{residue}{Choose to keep or remove the residues. If `removed` ptms will be separated by a hyphen.}
}
\value{
renamed PTM string
}
\description{
It shorthands PTM names into pretty annotations
}
\examples{
ptm_beautify(PTM = "TMAyl_correct (Any N-term); Butyryl (K28); Trimethyl (K37); Propionyl (K38)",
 lookup = histptm_lookup,
 software = "Proline", residue = 'keep')

 ptm_beautify(PTM = "TMAyl_correct (Any N-term); Butyryl (K28); Trimethyl (K37); Propionyl (K38)",
 lookup = histptm_lookup,
 software = "Proline", residue = 'remove')

ptm_beautify(PTM = "K[+126.06808]SAPATGGVK[+56.026215]K[+56.026215]PHR",
lookup = shorthistptm_mass,
software = "Skyline", residue = 'keep')

}
