% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_flagDupes.R
\name{ptm_flagDupes}
\alias{ptm_flagDupes}
\title{Flagging suspicious sequence-PTM combinations}
\usage{
ptm_flagDupes(df, PTM_col, var_col, select_cols)
}
\arguments{
\item{df}{A \code{dataframe}.}

\item{PTM_col}{PTM column name in the dataset.}

\item{var_col}{Column name of a unique identifier (e.g. peptide sequence) in the dataset.}

\item{select_cols}{Either one of the select helper functions with proper argument
(e.g. \code{starts_with("WT_")}) or a vector \code{c()} containing the names of
 intensity/abundance columns to check for duplication.}
}
\value{
A named \code{list} of two \code{tibbles}.

\emph{unique} \cr
* a non-redundant version of \code{df} after:\cr
    - removing identical duplicate rows \cr
    - removing redundant rows sharing identical intensity values \cr
    - marking with asterisk (*) redundant PTMs in \code{PTM_col}.\cr
      Redundant observations are reported in \code{duplicate} (see below) with their frequency \code{n}.

\emph{duplicate} \cr
* A subset \code{tibble} of \code{df} containing all de-duplicated identifications sharing
 identical intensities in \strong{ALL} \code{select_cols} columns.

\emph{freq_table} \cr
* A table summarizing the duplicate counts and their frequencies in the dataset.
}
\description{
In Proline result output, some sequence-PTM combinations have identical intensity values in all samples.
Usually those are isobaric PTM-sequence combinations that are most likely false positive hits.
}
