% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_rename.R
\name{ptm_rename}
\alias{ptm_rename}
\title{ptm_rename}
\usage{
ptm_rename(PTM, lookup = histonePTM::histptm_lookup, software = "Proline")
}
\arguments{
\item{PTM}{The PTM string to be shorthanded}

\item{lookup}{The shorthanded names to replace the long PTM names (histptm_lookup is a named vector provided by default. The user can provide any named vector adapted to the PTM being used.)}

\item{software}{The software used for proteomic analysis (ex. Proline (Default), MaxQaunt, Skyline, Fragpipe, ...)}
}
\value{
renamed PTM string
}
\description{
It shorthands PTM names into pretty annotations
}
\examples{
ptm_rename(PTM = "TMAyl_correct (Any N-term); Butyryl (K28); Trimethyl (K37); Propionyl (K38)",
 lookup = histptm_lookup,
 software = "Proline")


}
