% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_Uniprot.R
\name{ptm_Uniprot}
\alias{ptm_Uniprot}
\title{Retreive MS-reported PTMs from Uniprot}
\usage{
ptm_Uniprot(Uniprot_accession, save_file = FALSE)
}
\arguments{
\item{Uniprot_accession}{Uniprot protein-specific accession code.}

\item{save_file}{\code{logical}. save results as a \code{.csv} file (optional).}
}
\value{
A \code{tibble} and \code{.csv} file.
}
\description{
Get PTM information  reported in literature and documented by Uniprot.
}
\examples{
ptm_Uniprot(Uniprot_accession = "B9FXV5", save_file = TRUE)

}
