% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_manIsoCorreciton.R
\name{quant_manIsoCorreciton}
\alias{quant_manIsoCorreciton}
\title{Correct of Naturally Occuring Abundances}
\usage{
quant_manIsoCorreciton(df, time_col, t0, cond_col, avg_by)
}
\arguments{
\item{df}{A dataframe as PICor input.}

\item{time_col}{The time column.}

\item{t0}{Value of the reference sample in the time column ( e.g. "0" , "t0", "T0" ...)}

\item{cond_col}{Condition column (Concentration, disease vs non-disease ...)}

\item{avg_by}{(optional) A column to average by (e.g.TechReplicate).}
}
\value{
A dataframe with corrected and normalized values.
}
\description{
Subtract naturally occuring abundances of MS XICs in stable isotope tracing experiments
}
