% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_relIntensity.R
\name{quant_relIntensity}
\alias{quant_relIntensity}
\title{Normalize Intensity values}
\usage{
quant_relIntensity(
  df,
  select_cols,
  grouping_var,
  norm_method = c("peptide_family", "peptide_total")
)
}
\arguments{
\item{df}{A \code{tibble} or a \code{dataframe}}

\item{select_cols}{Either one of the select helper functions with proper argument
(e.g. \code{starts_with("WT_")}) or a vector \code{c()}containing the names of
 intensity/abundance columns to be normalized.}

\item{grouping_var}{A unqiue variable to group by like 'peptide sequence' or so. This mandatory if `norm_method` is 'peptide_family".}

\item{norm_method}{Normalization method. Either by 'peptide_family' or by 'peptide_total' intensity. The latter depends on what is in your dataset and if you have prefiltered it or not.}
}
\value{
The provided \code{dataframe} is returned with relative instead of absolute
intensity values of the abundance/intensity columns passed to the function.

\deqn{\text{norm_peptide_family} = \frac{\text{Intensity}_{\text{peptide X}}}{\sum \text{Intensity}_{\text{all peptides of the same sequence as X}}}}
\deqn{\text{norm_peptide_family} = \frac{\text{Intensity}_{\text{peptide X}}}{\sum \text{Intensity}_{\text{all peptides in the sample}}}}
}
\description{
A function to normalize intensities of peptidoforms by dividing an intensity value by the
sum of all intensities other peptidoforms of the same sequence.
}
\examples{
quant_relIntensity(iris, contains("Length"), grouping_var = "Species")
quant_relIntensity(iris, starts_with("Sepal"), grouping_var = "Species")

}
