% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_AAcountplot.R
\name{seq_AAcountplot}
\alias{seq_AAcountplot}
\title{Count and plot amino aicd occurence}
\usage{
seq_AAcountplot(..., plot = FALSE)
}
\arguments{
\item{...}{Protein sequence as a character string.}

\item{plot}{A \code{logical} argument to decide whether to plot or not the resuls.}
}
\value{
dataframe with 3 columns: name of a sequence, its amino acid residues and their count.
a plot if \code{plot} argument is set to \code{TRUE} .
}
\description{
Count the occurence of amino aicd residues in a protein sequence and plot their relative percentages.
}
\examples{
H3_seq <-  "ARTKQTARKSTGGKAPRKQLATKAARKSAPATGGVKKPHRYRPGTVALREIRRYQKSTELLIRKLPFQRL
           VREIAQDFKTDLRFQSSAVMALQEACEAYLVGLFEDTNLCAIHAKRVTIMPKDIQLARRIRGERA"
H4_seq <-  "SGRGKGGKGLGKGGAKRHRKVLRDNIQGITKPAIRRLARRGGVKRISGLIYEETRGVLKVFLENVIRDAV
           TYTEHAKRKTVTAMDVVYALKRQGRTLYGFGG"
seq_AAcountplot(H3_seq)
seq_AAcountplot(H3_seq, H4_seq, plot = TRUE)
}
