% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_spreadIntoSheets.R
\name{seq_spreadIntoSheets}
\alias{seq_spreadIntoSheets}
\title{Separate H3 sequence and sequence variants
Each sequence/variant will be saved to a separate excel worksheet}
\usage{
seq_spreadIntoSheets(
  df,
  split_by = "variant",
  histoneProtein = c("H3-H4", "H3", "H4"),
  label_type = c("labelbyK", "label"),
  FileName = ""
)
}
\arguments{
\item{df}{A dataframe}

\item{split_by}{name of the column containing sequence labels. The values will be the name of the excel sheets.}

\item{histoneProtein}{Three options H3-H4', 'H3' or 'H4'. In none is selected 'H3-H4' will be selected by default.}

\item{label_type}{label type to match by from \code{SequenceDB}. Two options 'labelbyK' (default) and 'label'.}

\item{FileName}{In case you want to save the file as \code{.xlsx}, provide a name with this}
}
\value{
A list of \code{Workbook} object(s), each with multiple excel worksheets.
}
\description{
Separate H3 sequence and sequence variants
Each sequence/variant will be saved to a separate excel worksheet
}
