% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sequenceDB.R
\docType{data}
\name{sequenceDB}
\alias{sequenceDB}
\title{Histone protein and peptide sequence labeling}
\format{
a named character vector of proteins.
}
\source{
in-house user defined names
}
\usage{
sequenceDB
}
\description{
a named list of proteins including a nested list of sequence, enzyme fragments and labels

\itemize{
  \item Protein. H3, H4, H2A, H2B and iRT
     \item seq. Full amino acid sequence of the core protein and its variants.
     \item Fragments. argC_frags/tryp_frags.
     \item label. Representation of fragments by position in sequence (e.g. "H3 (3-8)" )

}
}
\examples{
# sequenceDB[['H3']]
# sequenceDB[['H4']][['label']]
# sequenceDB$H3$argC_frags
# sequenceDB$H3$seq





}
\keyword{datasets}
