% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shorthistptm_mass.R
\docType{data}
\name{shorthistptm_mass}
\alias{shorthistptm_mass}
\title{Monoisotopic masse shifts and their corresponding PTMs}
\format{
a named character vector.
}
\source{
in-house
}
\usage{
shorthistptm_mass
}
\description{
a named vector containing masses of (labeled)-ptms and their shorthand names (inverse of histptm_lookup data)
}
\keyword{datasets}
