% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeHistone.R
\name{analyzeHistone}
\alias{analyzeHistone}
\title{Histone Proteomics Data Analysis}
\usage{
analyzeHistone(
  analysisfile,
  metafile,
  hist_prot,
  NA_threshold,
  output_result = c("single", "multiple")
)
}
\arguments{
\item{analysisfile}{`Proline` output excel file}

\item{metafile}{An excel file containing user-defined `SampleName` and `file` columns ('Condition', 'Bioreplicate' or 'TechReplicate' are `optional``).}

\item{hist_prot}{One of 4 histone proteins ('H3', 'H4', 'H2A', 'H2B'). If you to analyze them all use "All".}

\item{NA_threshold}{A filter value below which an identification having this value of missing intensity value(s) or more is discarded.}

\item{output_result}{Either `signle` or `multiple`. This will decided if all ids from different proteins are in one file (`single`) or in a separate file (`multiple`).}
}
\value{
excel files
}
\description{
Analyze histone bottom-up proteomics DDA data validated and quantified by `Proline` software.
}
