% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litRicher.R
\name{litRicher}
\alias{litRicher}
\title{Quotes from the histone literature}
\usage{
litRicher(quotes_num = 1)
}
\arguments{
\item{quotes_num}{Number of quotes to print out (\code{integer}). To read all the quotes, use "All".}
}
\value{
A character string containing a quote about histones and histone PTMs.
}
\description{
Print quote(s) cherry-picked from the histone literature.
}
\examples{
litRicher()
}
