% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_checkFileNames.R
\name{misc_checkFileNames}
\alias{misc_checkFileNames}
\title{Check match betwen two columns}
\usage{
misc_checkFileNames(raw_df, user_df, common_col)
}
\arguments{
\item{raw_df}{Dataframe containing acolumn of correct file names}

\item{user_df}{Dataframe containing a column of file names provided by the user to which another custom name are assigned}

\item{common_col}{column of file names in both dataframes (should be the same name in both).}
}
\value{
error or success message.
}
\description{
Could be used to check for user input file names if they match file names extracted from 'Search settings and infos' sheet of Proline export.
}
\examples{
df1 <- data.frame(file= c('HF1_022210', 'HF1_022222'))
df2 <- df1
misc_checkFileNames(df1, df2, common_col = 'file')
#>The provided and actual names match perfectly. Good job!✔
}
