% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_clearLabeling.R
\name{misc_clearLabeling}
\alias{misc_clearLabeling}
\title{Reverse PTM Labeling}
\usage{
misc_clearLabeling(ptm_string, rules, residue = "keep")
}
\arguments{
\item{ptm_string}{PTM-containing string}

\item{rules}{A named vector containing \code{(labeled_ptm = unlabeled_ptm)}. Example: \code{"pr" = "un"}.}

\item{residue}{Specify wether to remove residue(capitlized letter followed by number). Either `keep` or `remove`.}
}
\value{
unlabeled string
}
\description{
For final representation it is better to replace chemical labeling with biological state. For example, replace 'propionyl' with 'unmodified'
or 'butyryl' with 'me1' (since me1+pr = bu).
}
