% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeNplot.R
\name{normalizeNplot}
\alias{normalizeNplot}
\title{Normalize and Plot}
\usage{
normalizeNplot(
  df_corrected,
  df_meta,
  ptm_col,
  seq_col,
  seq_stretch_col,
  int_cols,
  save_file = TRUE,
  labeling = c("none", "PA", "TMA", "PIC_PA"),
  rules = NULL,
  plot_title = NULL,
  fun = "mean",
  error_type = NULL,
  conf_level = 0.95,
  cond_order = NULL,
  save_plot = TRUE
)
}
\arguments{
\item{df_corrected}{A dataframe containing the corrected raw intensity values along with the necessary column (see below)}

\item{df_meta}{A dataframe containing at least the 'Condition' and 'SampleName'. The latter MUST be the same the names of the `int_cols`.}

\item{ptm_col}{A column containing PTM. The function `misc_clearLabeling()` according to `labeling` will be applied to remove any labeling.}

\item{seq_col}{A column containing stripped sequences. Will be used for normalization. Could also be used as `plot_title`.}

\item{seq_stretch_col}{A column containing the label of the sequences (will be used as title of the y axis).}

\item{int_cols}{Intensity columns. `tidyselect` functions can be used (e.g. starts_with('abundance_')) or character vector (e.g `c('col1', 'col2')`).}

\item{save_file}{(\code{bool}) TRUE (default) to save the normalized file.}

\item{labeling}{Labeling method (e.g. 'PA', 'TMA', 'PA_PIC', or 'none' (default)). Check `misc_clearLabeling()`.}

\item{rules}{(optional) Rules to unlabel PTM strings. Check `misc_clearLabeling()`.}

\item{plot_title}{A string or a column name.}

\item{fun}{A statistical test ('mean' or 'median'). This will be the height of the bars.}

\item{error_type}{(optional) Argument to add error bars. It takes either "CI", "SD" or "SE" (check `plot_jitterbarIntvsPTM()`).}

\item{conf_level}{(optional) The confidence level (0.95 (default)). Only useful if you want to add error bars (see `error_type`).}

\item{cond_order}{(optional) A character vector containing the conditions in the preferred order. If you want a specific order of the condition.}

\item{save_plot}{(\code{bool}) TRUE (default) to save the plot(s). Otherwise, set it to FALSE.}
}
\value{
Same output as `plot_jitterbarIntvsPTM()`. A dataframe grouped by 'id_col' (and 'plot_title' if passed) with a nested list-column harboring the generated
bar plots (representing either 'mean' or 'median') with jitter points (corresponding to individual measurements per 'condition').
}
\description{
There are cases where the output of `analyzeHistone()` function is not satisfactory. This happens when the software integrates the same peak for two nearly coeluting peptidoforms
as K18acK23un and K18unK23un. A need to re-normalize and to re-plot the data is the reason this function exists.
}
