% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotateSpectrum.R
\name{plot_annotateSpectrum}
\alias{plot_annotateSpectrum}
\title{Annotate MS/MS Spectra}
\usage{
plot_annotateSpectrum(
  Profrma_peptide,
  prec_mz,
  prec_z,
  mz,
  intensity,
  title = "",
  annot_diag = FALSE,
  nloss = TRUE,
  min_int = 0.01,
  ion_types = "by",
  tol_mz = "20",
  tol_mode = "ppm",
  output_plot_name = ""
)
}
\arguments{
\item{Profrma_peptide}{Peptide in proforma representation}

\item{prec_mz}{m/z of the precursor ion (double)}

\item{prec_z}{charge state of the precursor ion (integer)}

\item{mz}{\code{numeric} vector containing mz values}

\item{intensity}{\code{numeric} vector containing intensity values}

\item{title}{title of the annotate MS/MS plot}

\item{annot_diag}{(bool, FALSE (default)) Annotate CyIm ions produced from modified lysines.}

\item{nloss}{(bool, TRUE (default)). Annotate neutral losses (-NH3 by "*" and -H2O by "o")}

\item{min_int}{(double, 0.01 (default)). To filter intensity which are `min_int` of the base peak.}

\item{ion_types}{Fragment ion types (i.e. "by' (Default))}

\item{tol_mz}{fragment mass tolerance (e.g, 20)}

\item{tol_mode}{'ppm' (Default) or 'Da'}

\item{output_plot_name}{Name of the plot in case you want to save it}
}
\value{
Annotated MS/MS plot
}
\description{
wrapper function around functions from \code{spectrum_utils} library in Python.
}
