% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotateSpectrum.R
\name{plot_annotateSpectrum}
\alias{plot_annotateSpectrum}
\title{Annotate MS/MS Spectra}
\usage{
plot_annotateSpectrum(Profrma_peptide, prec_mz, prec_z, mz, intensity, title)
}
\arguments{
\item{Profrma_peptide}{Peptide in proforma representation}

\item{prec_mz}{m/z of the precursor ion}

\item{prec_z}{charge state of the precursor ion}

\item{mz}{\code{numeric} vector containing mz values}

\item{intensity}{\code{numeric} vector containing intensity values}

\item{title}{title of the annotate MS/MS plot}
}
\value{
Annotate MS/MS plot
}
\description{
wrapper function around functions from \code{spectrum_utils} library in Python.
}
