% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CVs.R
\name{plot_CVs}
\alias{plot_CVs}
\title{Viusalizaing Coefficients of Variation (CVs)}
\usage{
plot_CVs(
  df,
  cond_col,
  cv_col,
  scale = 100,
  save_plot = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{cond_col}{The condition column.}

\item{cv_col}{The column containing the coefficients of variations}

\item{scale}{1 or 100. If CVs are in percentage, use 1. If not, set it to 100.}

\item{save_plot}{Save plot (bool; FALSE (default))}

\item{output_dir}{Folder to save the plot in. Defualts to working directory.}
}
\value{
list of two ggplot2. the violin_plot and bar_plot of CVs
}
\description{
Plot distribution of CVs as violin plot and count IDs that have CV of 20 % and 5 % or less.
}
