% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intDensity.R
\name{plot_intDensity}
\alias{plot_intDensity}
\title{A ridgeline visualization of intensities per sample}
\usage{
plot_intDensity(df, int_col, save_plot = "")
}
\arguments{
\item{df}{Dataset in wide format.}

\item{int_col}{The intensity columns. Either a vector `c('column1', 'column2', ...)` or using tidyselect functions}

\item{save_plot}{plot_name.(png, svg, ...)
like `starts_with('abundant_')` or `contains('intensity_')`}
}
\value{
ridgeline plot
}
\description{
A plot the gives an eye-bird view of the intensity distribution in each sample.
}
