% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jitterbarIntvsPTM.R
\name{plot_jitterbarIntvsPTM}
\alias{plot_jitterbarIntvsPTM}
\title{Intensity vs PTM grouped bar plot with jittered points}
\usage{
plot_jitterbarIntvsPTM(
  dataset,
  x_axis,
  y_axis,
  condition,
  sequence,
  label = "",
  plot_title = "",
  fun = c("median", "mean"),
  scale = 100,
  save_plot = FALSE
)
}
\arguments{
\item{dataset}{your data loaded as dataframe. Must be in long format.}

\item{x_axis}{x variable (PTM column)}

\item{y_axis}{y variable (intensity column). If already in %, set \code{scale} to 1.}

\item{condition}{The condition column (WT vs disease, concentration, ...)}

\item{sequence}{peptide sequence column}

\item{label}{peptide label to be used for y-axis title (e.g. 'H3.3')}

\item{fun}{median (default) or mean. This will be the height of the bar.}

\item{scale}{100 (default). If you want to keep values as they are use 1.}

\item{save_plot}{(\code{logical})}

\item{title_plot}{A plot title (optional)}
}
\value{
ggplot object
}
\description{
Each individual measurement will be plotted as datapoint.
Their median/mean will be represented as a bar.
Each bar will be colored according to the condition.
}
