% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_labelingAssessment.R
\name{ptm_labelingAssessment}
\alias{ptm_labelingAssessment}
\title{Assessing (over)Labeling Effeciency}
\usage{
ptm_labelingAssessment(df, seq_col, seq, ptm_col, int_col, plot_title = "")
}
\arguments{
\item{df}{Your dataset}

\item{seq_col}{column containing bare peptide sequences}

\item{seq}{sequence to base the assessment on it}

\item{ptm_col}{column containing peptide ptm information (for overTMA: TMAyl, tmaNt, and tmame1, tma;, for overProp: Propionyl, prNt, pr). Only
two representations of PTMs can be recognized to filter upon: Propionly (K12) or K27bu-S32pr ....)}

\item{int_col}{sample columns containing intensity values}

\item{plot_title}{A title for the plot}
}
\value{
list of two dataframes (raw and normalized values) and a bar plot.
}
\description{
Some time overlabeling is very abundant which decrease sensitivity and affect quantification accuracy.
}
