% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_siteAbundance.R
\name{ptm_siteAbundance}
\alias{ptm_siteAbundance}
\title{Site-specific abundance
Calculates the abundance of an individual histone PTM site.}
\usage{
ptm_siteAbundance(
  df,
  metadata,
  ptm_col,
  id_col,
  int_cols,
  format = c("wide", "long"),
  remove_ptm = NULL,
  save_plot = FALSE
)
}
\arguments{
\item{df}{Dataframe containing normalized values}

\item{ptm_col}{PTM column containing PTMs in Brno nomenclature, separated by "-" if multiple PTMs exist (e.g. 'K27ac-K36bu')}

\item{id_col}{The identity column(s) (e.g. column containing sequences).}

\item{int_cols}{Column(s) containing intensity values.}

\item{format}{Format of `df` ('wide' (default) or 'long'))}

\item{remove_ptm}{A ptm to be excluded. provided as a string.}

\item{save_plot}{\code{bool, FALSE} (default)}

\item{df_meta}{Dataframe containing the sample design. Must contain the column "SampleName" and "Condition".}
}
\value{
dataframe with id_col(s), `PTMsite` containing the quantified sites, and the intensity column(s)
}
\description{
Site-specific abundance
Calculates the abundance of an individual histone PTM site.
}
