% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_toProForma.R
\name{ptm_toProforma}
\alias{ptm_toProforma}
\title{Proforma modified peptide representation}
\usage{
ptm_toProforma(seq, mod, Nterm = FALSE)
}
\arguments{
\item{seq}{The sequence to be modified.}

\item{mod}{The modifications with their positions on the peptide}

\item{Nterm}{N-terminal modification}
}
\value{
sequence with monoisotopic mass of ptm at its modification site.
}
\description{
Proforma modified peptide representation
}
\examples{
ptm_toProforma(seq = "KSAPATGGVKKPHR",
               mod = "Propionyl (Any N-term); Lactyl (K1); Dimethyl (K10); Propionyl (K11)")

ptm_toProforma(  seq = "KQLATKVAR",
                 mod = "Propionyl (Any N-term); Propionyl (K1); Propionyl (K6)",
               Nterm = "[+56.026]-")

}
