% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_toProForma.R
\name{ptm_toProForma}
\alias{ptm_toProForma}
\title{Convert sequence-PTM to ProForma notation.}
\usage{
ptm_toProForma(seq, mod, Nterm = FALSE)
}
\arguments{
\item{seq}{The sequence to be modified.}

\item{mod}{The modifications with their positions on the *peptide*.}

\item{Nterm}{N-terminal modification}
}
\value{
character string
}
\description{
Convert Proline modification string to ProForma notation.
}
\examples{
ptm_toProForma(seq = "KSAPATGGVKKPHR",
               mod = "Propionyl (Any N-term); Lactyl (K1); Dimethyl (K10); Propionyl (K11)")

ptm_toProForma(  seq = "KQLATKVAR",
                 mod = "Propionyl (Any N-term); Propionyl (K1); Propionyl (K6)",
               Nterm = "[+56.026]")

}
