% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_Uniprot.R
\name{ptm_Uniprot}
\alias{ptm_Uniprot}
\title{Retreive proteomics-validated PTMs from Uniprot}
\usage{
ptm_Uniprot(accession, config = httr::accept("application/json"), ...)
}
\arguments{
\item{accession}{The accession number of the interrogated protein.}

\item{config}{For advanced users.}

\item{...}{for any later addition.}
}
\value{
A \code{tibble} with PTMs reported in Uniprot with their labels (gold, silver, ...).
Note not all PTMs are yet documented in Uniport.
}
\description{
A function that access to Uniprot to retreive documented and classifed PTMs along with
many other features such as universal spectrum idetifier (USI) if reported by the database of origin.
}
\examples{
ptm_Uniprot(acession = "B9FXV5")
}
