% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_coefVariation.R
\name{quant_coefVariation}
\alias{quant_coefVariation}
\title{Calculate Coefficient of Variation}
\usage{
quant_coefVariation(
  df,
  df_meta,
  int_col,
  seq_col,
  ptm_col,
  format = c("wide", "long"),
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{df_meta}{dataframe containing at least these two columns: `SampleName` and `Condition`.
If `TechnicalReplicate` exist and needed to be treated separately, the easiest way to transform your `df` into long format,
then use `format = long` and add replace the `...` argument with the `TechnicalReplicate` column name.}

\item{int_col}{intensity columns}

\item{seq_col}{sequence column}

\item{ptm_col}{PTM column}

\item{format}{'wide' or 'long'}

\item{...}{to add any other columns}
}
\value{
The input dataframe plus 3 columns per condition: `sd_condition`, `avg_condition` and `CV_condition`.
}
\description{
Calculate Coefficient of Variation
}
