% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_me1LabelingPlot.R
\name{quant_me1LabelingPlot}
\alias{quant_me1LabelingPlot}
\title{Assessing Monomethyl Derivatization}
\usage{
quant_me1LabelingPlot(
  df,
  seq_col,
  ptm_col,
  int_col,
  cond_col = NULL,
  me1 = "me1",
  me1_label = "bu",
  format = c("wide", "long"),
  df_meta,
  save_plot = FALSE
)
}
\arguments{
\item{df}{Dataframe containing proteomic analysis dataset}

\item{seq_col}{Sequence column}

\item{ptm_col}{PTM column}

\item{int_col}{Intensity column(s)}

\item{cond_col}{(`optional`) Condition column (e.g. concentration) used for legend. If absent will be replaced by "None".}

\item{me1}{Monomethyl representation inside the `ptm_col`column (i.e. 'me1' (`default`) or 'Methyl', ...)}

\item{me1_label}{Monomethyl+label representation inside the `ptm_col`column (i.e. 'bu' (`default`) or 'Butyryl, ....)}

\item{format}{if set to `wide` (most often), data will be reshaped into long format before plotting. Check `df_meta`.}

\item{df_meta}{(`optional`). If df is in `wide` format, then one can provide df_meta containing at least two columns (`SampleName` and `Condition`).
`SampleNames` MUST match exactly the names of the `int_col`columns.}

\item{save_plot}{(`logical`). If `TRUE`, all the plots will be saved in one pdf. If `FALSE` (`default`), the plot(s) will be shown within the Plots pane.)}
}
\value{
Boxplot with jittered points to assess me1 labeling when using anhydrides to label lysines.
}
\description{
Evaluate the efficiency of monomethylated lysine derivarization per sequencexPTM combination.
}
