% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_getHistPeptides.R
\name{seq_getHistPeptide}
\alias{seq_getHistPeptide}
\title{Extract selected histone peptides}
\usage{
seq_getHistPeptide(
  df,
  seq_col = sequence,
  histoneProtein = c("All", "H3", "H4", "H2A", "H2B", "H1")
)
}
\arguments{
\item{df}{A \code{dataframe} containing the proteomics experiment results.}

\item{seq_col}{The name of the column containing peptide sequences.}

\item{histoneProtein}{Histone protein(s) to be studied: "H3", "H4" or both "H3-H4".}
}
\value{
A subset of \code{df} only containing peptides of the defined histone protein(s) with \href{https://biognosys.com/product/irt-kit/}{Biognosys}
iRTs if present.
}
\description{
Isolate specific fragments of user-defined histone protein(s).
}
